Reflect ConductR 2.0
We now refer what was ConductR 1.2 as 2.0 given the major changes in its topology in particular i.e. the distinction between core and agent.

I've also updated the main page's copy to reflect the why, what and how of ConductR.

Finally, I've updated the releases of everything. In particular I've migrated the code from Play 2.3 to Play 2.5, along with lots of updates to Akka 2.4, distributed data and so forth. This had to be done at some point. We could go further here in terms of retiring iteratees altogether, but that can be for another day. These updates take care of the bulk of modernisation and will last for some time into the future.